Corregido el init.

+++ Init.php
@@ -25,8 +25,12 @@ class Init extends InitClass
         // al actualizar a la versión 3.0
         $dataBase = new DataBase();
         $columns = ['nombre'];
-        $tables = ['employee_contracts', 'employee_attendance_management_yn', 'drivers', 'helpers', 'collaborators'];
+        $tables = ['employee_contracts', 'employees_attendance_management_yn', 'drivers', 'helpers', 'collaborators'];
         foreach ($tables as $table) {
+            // preguntamos si exite la tabla
+            if (false === $dataBase->tableExists($table)) {
+                continue;
+            }
             foreach ($dataBase->getColumns($table) as $column) {
                 if (in_array($column->name, $columns)) {
                     $sql = 'ALTER TABLE ' . $table . ' DROP COLUMN ' . $column->name;
